feat: implement step timeout part 1
