+++ pkg/runner/runner.go
@@ -3,14 +3,12 @@ package runner
 import (
 	"context"
 	"fmt"
-	"path/filepath"
 	"sync/atomic"
 	"testing"
 
 	"github.com/kyverno/chainsaw/pkg/apis/v1alpha1"
 	"github.com/kyverno/chainsaw/pkg/client"
 	"github.com/kyverno/chainsaw/pkg/discovery"
-	"github.com/kyverno/chainsaw/pkg/resource"
 	runnerclient "github.com/kyverno/chainsaw/pkg/runner/client"
 	"github.com/kyverno/chainsaw/pkg/runner/logging"
 	"github.com/kyverno/chainsaw/pkg/runner/namespacer"
@@ -67,11 +65,13 @@ func Run(cfg *rest.Config, config v1alpha1.ConfigurationSpec, tests ...discovery
 				t.Helper()
 				t.Parallel()
 				ctx := Context{
+					config:     config,
+					ctx:        context.Background(),
+					namespacer: nspacer,
 					clientFactory: func(t *testing.T, logger logging.Logger) client.Client {
 						t.Helper()
 						return runnerclient.New(t, logger, c, !config.SkipDelete)
 					},
-					namespacer: nspacer,
 				}
 				t.Cleanup(func() {
 					if t.Skipped() {
@@ -106,84 +106,5 @@ func runTest(t *testing.T, ctx Context, test discovery.Test) {
 	for i := range test.Spec.Steps {
 		step := test.Spec.Steps[i]
 		executeStep(t, logging.NewStepLogger(t, fmt.Sprintf("step-%d", i+1)), ctx, test.BasePath, step)
-	}
-}
-
-func executeStep(t *testing.T, logger logging.Logger, ctx Context, basePath string, step v1alpha1.TestStepSpec) {
-	t.Helper()
-	c := ctx.clientFactory(t, logger)
-	// Delete the Objects before the test step is executed
-	for _, delete := range step.Delete {
-		// Use your dynamic listing logic if the name is not provided
-		if delete.Name == "" {
-			u, err := client.ListResourcesToDelete(c, delete)
-			if err != nil {
-				t.Fatal(err)
-			}
-			for _, item := range u.Items {
-				currentItem := item
-				t.Logf("=== DELETE %s/%s", delete.APIVersion, delete.Kind)
-				if err := client.DeleteResource(context.TODO(), c, &currentItem); err != nil {
-					t.Fatal(err)
-				}
-			}
-		} else {
-			resource := client.NewResource(delete.APIVersion, delete.Kind, delete.Name, delete.Namespace)
-			t.Logf("=== DELETE %s/%s", delete.APIVersion, delete.Kind)
-			if err := client.DeleteResource(context.TODO(), c, resource); err != nil {
-				t.Fatal(err)
-			}
-		}
-	}
-	for _, apply := range step.Apply {
-		resources, err := resource.Load(filepath.Join(basePath, apply.File))
-		if err != nil {
-			t.Fatal(err)
-		}
-		for i := range resources {
-			resource := &resources[i]
-			if err := ctx.namespacer.Apply(resource); err != nil {
-				t.Fatal(err)
-			}
-			logging.ResourceOp(logger, "APPLY", client.ObjectKey(resource), resource)
-			err := client.CreateOrUpdate(context.Background(), c, resource)
-			if err != nil {
-				t.Fatal(err)
-			}
-		}
-	}
-	for _, assert := range step.Assert {
-		resources, err := resource.Load(filepath.Join(basePath, assert.File))
-		if err != nil {
-			t.Fatal(err)
-		}
-		for i := range resources {
-			resource := &resources[i]
-			if err := ctx.namespacer.Apply(resource); err != nil {
-				t.Fatal(err)
-			}
-			logging.ResourceOp(logger, "ASSERT", client.ObjectKey(resource), resource)
-			if err := client.Assert(context.Background(), resources[i], c); err != nil {
-				t.Fatal(err)
-			}
-		}
-	}
-	for _, e := range step.Error {
-		resources, err := resource.Load(filepath.Join(basePath, e.File))
-		if err != nil {
-			t.Fatal(err)
-		}
-		for i := range resources {
-			resource := &resources[i]
-			if err := ctx.namespacer.Apply(resource); err != nil {
-				t.Fatal(err)
-			}
-			logging.ResourceOp(logger, "ERROR", client.ObjectKey(resource), resource)
-			// Using the Error function to handle the error assertion
-			err := client.Error(context.Background(), resources[i], c)
-			if err != nil {
-				t.Fatal(err)
-			}
-		}
 	}
 }
